+++ module/core/derive_tools/tests/inc/deref/generics_types.rs
@@ -2,7 +2,7 @@ use core::ops::Deref;
 use derive_tools::Deref;
 
 #[ allow( dead_code ) ]
-// #[ derive( Deref ) ]
+#[ derive( Deref ) ]
 struct GenericsTypes< T >( T );
 
 include!( "./only_test/generics_types.rs" );